fix(enum) : 2023. 11 Survivor enum 타입을 int로 캐스팅하는 것을 일단 변경

+++ 0_Scripts/Project/Survival/00_Select/0_Base/SelectedGroup.cs
@@ -47,7 +47,7 @@ public class SelectedGroup
             for (int i = 0; i < curSelected.Count; i++)
             {
 
-                if (curSelected[i].MyState != (int)STATE_SELECTABLE.BUILDING_UNFINISHED)
+                if (curSelected[i].MyState != STATE_SELECTABLE.BUILDING_UNFINISHED)
                 {
 
                     return false;
